+++ collector.c
@@ -44,6 +44,7 @@ collector_runner(void *s)
 	size_t totticks;
 	size_t totmetrics;
 	size_t totblackholes;
+	size_t totdiscards;
 	size_t totqueued;
 	size_t totstalls;
 	size_t totdropped;
@@ -51,6 +52,7 @@ collector_runner(void *s)
 	size_t ticks;
 	size_t metrics;
 	size_t blackholes;
+	size_t discards;
 	size_t queued;
 	size_t stalls;
 	size_t dropped;
@@ -75,6 +77,7 @@ collector_runner(void *s)
 	size_t (*d_ticks)(dispatcher *) = NULL;
 	size_t (*d_metrics)(dispatcher *) = NULL;
 	size_t (*d_blackholes)(dispatcher *) = NULL;
+	size_t (*d_discards)(dispatcher *) = NULL;
 	size_t (*d_sleeps)(dispatcher *) = NULL;
 	size_t (*a_received)(aggregator *) = NULL;
 	size_t (*a_sent)(aggregator *) = NULL;
@@ -119,6 +122,7 @@ collector_runner(void *s)
 				d_ticks = dispatch_get_ticks_sub;
 				d_metrics = dispatch_get_metrics_sub;
 				d_blackholes = dispatch_get_blackholes_sub;
+				d_discards = dispatch_get_discards_sub;
 				d_sleeps = dispatch_get_sleeps_sub;
 				a_received = aggregator_get_received_sub;
 				a_sent = aggregator_get_sent_sub;
@@ -131,6 +135,7 @@ collector_runner(void *s)
 				d_ticks = dispatch_get_ticks;
 				d_metrics = dispatch_get_metrics;
 				d_blackholes = dispatch_get_blackholes;
+				d_discards = dispatch_get_discards;
 				d_sleeps = dispatch_get_sleeps;
 				a_received = aggregator_get_received;
 				a_sent = aggregator_get_sent;
@@ -156,17 +161,22 @@ collector_runner(void *s)
 		totticks = 0;
 		totmetrics = 0;
 		totblackholes = 0;
+		totdiscards = 0;
 		totsleeps = 0;
 		for (i = 0; dispatchers[i] != NULL; i++) {
 			totsleeps += sleeps = d_sleeps(dispatchers[i]);
 			totticks += ticks = d_ticks(dispatchers[i]);
 			totmetrics += metrics = d_metrics(dispatchers[i]);
 			totblackholes += blackholes = d_blackholes(dispatchers[i]);
+			totdiscards += discards = d_discards(dispatchers[i]);
 			snprintf(m, sizem, "dispatcher%d.metricsReceived %zu %zu\n",
 					i + 1, metrics, (size_t)now);
 			send(metric);
 			snprintf(m, sizem, "dispatcher%d.metricsBlackholed %zu %zu\n",
 					i + 1, blackholes, (size_t)now);
+			send(metric);
+			snprintf(m, sizem, "dispatcher%d.metricsDiscarded %zu %zu\n",
+					i + 1, discards, (size_t)now);
 			send(metric);
 			snprintf(m, sizem, "dispatcher%d.wallTime_us %zu %zu\n",
 					i + 1, ticks, (size_t)now);
@@ -180,6 +190,9 @@ collector_runner(void *s)
 		send(metric);
 		snprintf(m, sizem, "metricsBlackholed %zu %zu\n",
 				totblackholes, (size_t)now);
+		send(metric);
+		snprintf(m, sizem, "metricsDiscarded %zu %zu\n",
+				totdiscards, (size_t)now);
 		send(metric);
 		snprintf(m, sizem, "dispatch_wallTime_us %zu %zu\n",
 				totticks, (size_t)now);