+++ ChangeLog.md
@@ -5,6 +5,8 @@
 * **router** now emits the bad string if an invalid port was given
 * **router** new support for dynamic clusters
   [Issue #367](https://github.com/grobian/carbon-c-relay/issues/367)
+* **dispatcher** allow limits to input and metric lengths
+  [Issue #354](https://github.com/grobian/carbon-c-relay/issues/354)
 
 ### Bugfixes